Version 0.7
Changes to the constructor:
- passing in cookies was removed because I now realized those would
have been overwritten anyways and that's lame.
- Sessions are no longer created in the login method, instead they're
created in the constructor. You can now pass in a session object that
you've already created, and PyEdsby will use that.
- You can pass in your instance metadata, which keeps us from having to
scrape it and lets you make PyEdsby instances a little faster (more on
that in a moment)

New in 0.7: Instance Metadata!
As it turns out, Edsby instances themselves have NIDs, and these NIDs
are unique (Many thanks to @davidbgeek's eagle eye for discovering
this!). We have to scrape this data from the default page of your Edsby
instance because we can't retrieve it from anywhere else, so there are
a couple of new methods in 0.7 for this:

- parseInstanceMetadata, which performs the actual scraping/parsing of
the metadata dict from the base Edsby URL and returns it.
- getInstanceMetadata, which returns the instance metadata PyEdsby is
holding internally.

We are also keeping track of instance metadata in self.instanceMeta, so
any methods that were previously using the hardcoded NID of my instance
have now been updated appropriately.

Other misc updates:
setCookies now accepts cookies in both dict and cookie jar formats, and
actually sets cookies in the correct way. This change shouldn't break
anything, because it was broken in the first place and now it isn't.