Added more methods for groups
Added support for deleting group posts (replies + top level posts), posting files in group feeds, I plan on fixing class posting soon. Replaced getRawGroupData with a method that actually returns the raw data of a group, what used to be getRawGroupData is now getStudentGroups. Added a method that returns the moderators of a group. Made a minor change to getCalendarData, it now has a specified date.